requirements.txt
### development
absl_py==2.0.0
confluent_kafka==2.2.0
freezegun==1.2.2
gitpython==3.1.40
jsonpath_ng==1.4.3
PyDriller==2.5.1
python_dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
semantic_version==2.10.0
wheel==0.38.4

# json5 for vscode config file handling:
# https://pypi.org/project/json5/
# mainly for trailing comma and comments support
json5==0.9.11

### best practices
black==23.10.1
coverage==7.3.2
pre-commit==2.20.0

### flake8
flake8==6.1.0
# flake8-annotations==3.0.0 # ! this requires flake8 >= 5
# flake8-comprehensions
# flake8-bugbear
# flake8-builtins
# flake8-tuple
# flake8-use-fstring
isort==5.12.0
mypy==1.6.1
types-python-dateutil==2.8.19
types-requests==2.27.31
types-PyYAML==6.0.12.12
types-urllib3==1.26.25.14

# Please ensure the wheel version is the same as the version used in third_party/python/BUILD
# i.e. contracts_api-4.5.3-py3-none-any.whl is version 4.5.3
# You can install the package manually which can be done from
# https://docs.thoughtmachine.net/vault-core/4-5/EN/reference/contracts/development_and_testing/#contracts_sdk
# ! please note this S3 bucket is only accessible on TM wifi/VPN
# https://s3-eu-west-1.amazonaws.com/tm-py-wheels/py39/contracts_api-4.6.5-py3-none-any.whl
pytest==8.3.5